+++ floodfill.h
@@ -2,4 +2,4 @@
 #include "queue.h"
 
 //returns the length of the largest cluster
-extern int findLargestCluster(int** array, int size);
+extern int findLargestCluster(short** array, int size);
